Player - Enemy Collisions (To fix)
Collisions :  Does not collide properly. After the second or third collision decremens all life and dies (Without reason).
When collides with an enemy the second (And the rest of the enemies) start spawning at a higher velocity, dissapear and damages player (Without a proper collision, it just collide at distance)
Suggestion : Check for loops, enemy bools and movement (Like conditions or velocities)